requirements.txt
streamlit
anthropic
pdfplumber
python-docx
openpyxl
Pillow
